% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_seats_dhondt.R
\name{allocate_seats_dhondt}
\alias{allocate_seats_dhondt}
\title{Calculates the allocation of seats to political parties in legislative
elections using the D'Hondt method.}
\usage{
allocate_seats_dhondt(
  parties_votes,
  total_electors_or_votes,
  seats_to_allocate,
  votes_threshold = 0
)
}
\arguments{
\item{parties_votes}{numeric vector with the votes of each political party.
Must be named with the names of the political parties.}

\item{total_electors_or_votes}{integer value of the total amount of
registered electors or total amount of votes used to set the minimum
threshold.}

\item{seats_to_allocate}{integer value of the total amount of seats to be
allocated in the election.}

\item{votes_threshold}{double value, between 0 and 1, to represent the
minimum percentage of the total electors or votes needed to enter in the
allocation.}
}
\value{
An integer vector.
}
\description{
Calculates the allocation of seats to political parties in legislative
elections using the D'Hondt method.
}
\examples{
allocate_seats_dhondt(
  parties_votes = c(
    "Party A" = 3550321L,
    "Party B" = 3444446L,
    "Party C" = 669865L,
    "Party D" = 609158L,
    "Party E" = 389295L,
    "Party F" = 263515L
  ),
  total_electors_or_votes = 12888231L,
  seats_to_allocate = 35,
  votes_threshold = 0.03
)
}
